Convert ReportSearch to TS, consolidate regex
